pyproject.toml
[project]
name = "socketio_pg_manager"
version = "0.0.2"
authors = [
    { name = "Chenyu Wang", email = "wangchenyu2017@gmail.com" },
]
description = "Postgres client manager for python-socketio"
classifiers = [
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "python-socketio >= 5.11.0",
    "psycopg[binary] >= 3.1.17",
    "psycopg-pool >= 3.2.3",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/MonoidDev/socketio_pg_manager"
"Bug Tracker" = "https://github.com/MonoidDev/socketio_pg_manager/issues"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [
    "src",
]
namespaces = false

[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"
